CLOUDP-113301: Create atlascli packages tasks
