+++ wix/atlascli.json
@@ -1,0 +1,27 @@
+{
+  "product": "Atlas CLI",
+  "company": "MongoDB, Inc.",
+  "license": "LICENSE",
+  "upgrade-code": "a995cda5-9731-4158-910c-c8d15371dbda",
+  "files": {
+    "guid": "c7b0df57-c3c1-4591-becd-f0f73a912eb8",
+    "items": [
+      "bin/atlas.exe"
+    ]
+  },
+  "env": {
+    "guid": "94faac3d-4478-431c-8497-fba55dcfb249",
+    "vars": [
+      {
+        "name": "PATH",
+        "value": "[INSTALLDIR]",
+        "permanent": "yes",
+        "system": "no",
+        "action": "set",
+        "part": "last"
+      }
+    ]
+  },
+  "shortcuts": {},
+  "choco": {}
+}
